+++ src/scicloj/metamorph/ml/loss.clj
@@ -43,13 +43,18 @@ label-seq is a sequence of values.  The answer is considered correct
 if the key highest probability in the model output entry matches
 that label."
   ^double [lhs rhs]
-  (errors/when-not-errorf
-   (= (dtype/ecount lhs)
-      (dtype/ecount rhs))
-   "Ecounts do not match: %d %d"
-   (dtype/ecount lhs) (dtype/ecount rhs))
-  (/ (dtype/ecount (argops/binary-argfilter :tech.numerics/eq lhs rhs))
-     (dtype/ecount lhs)))
+  (let [lhs-all-numeric (every? number? lhs)
+        rhs-all-numeric (every? number? rhs)]
+    (errors/when-not-errorf (= lhs-all-numeric rhs-all-numeric)
+                            "lhs / rhs need to be either both numeric or both non-numeric: lhs: %s rhs: %s" lhs rhs)
+
+    (errors/when-not-errorf
+     (= (dtype/ecount lhs)
+        (dtype/ecount rhs))
+     "Ecounts do not match: %d %d"
+     (dtype/ecount lhs) (dtype/ecount rhs))
+    (/ (dtype/ecount (argops/binary-argfilter :tech.numerics/eq lhs rhs))
+       (dtype/ecount lhs))))
 
 
 (defn classification-loss