+++ src/scicloj/metamorph/ml/tools.clj
@@ -2,6 +2,14 @@
   (:require
    [clojure.pprint :as pprint]))
 
+(defn- maybe-dissoc [x k]
+  (if (associative? x)
+    (dissoc x k)
+    x))
+
+(defn- maybe-empty? [x]
+  (and (sequential? x) (empty? x)))
+
 (defn dissoc-in
   "Dissociate a value in a nested assocative structure, identified by a sequence
   of keys. Any collections left empty by the operation will be dissociated from
@@ -10,10 +18,10 @@
   (if-let [[k & ks] (seq ks)]
     (if (seq ks)
       (let [v (dissoc-in (get m k) ks)]
-        (if (empty? v)
-          (dissoc m k)
+        (if (maybe-empty? v)
+          (maybe-dissoc m k)
           (assoc m k v)))
-      (dissoc m k))
+      (maybe-dissoc m k))
     m))